fix: tests

+++ test/split.test.js
@@ -23,15 +23,17 @@ describe('File', function () {
     });
 
     afterEach(async function () {
-        fs.readdir(outDir, (err, files) => {
-            if (err) throw err;
+        if (fs.existsSync(outDir)) {
+            fs.readdir(outDir, (err, files) => {
+                if (err) throw err;
 
-            for (const file of files) {
-                fs.unlink(path.join(outDir, file), err => {
-                    if (err) throw err;
-                });
-            }
-        });
+                for (const file of files) {
+                    fs.unlink(path.join(outDir, file), err => {
+                        if (err) throw err;
+                    });
+                }
+            });
+        }
     });
 
     describe('empty', function () {
